+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Location is now updated on registration. Customers location is truncated to 2
   decimal places based on their postcode.
+* Location is also updated on changing a users postcode
 
 ## Bug Fixes